+++ apstools/utils/plot.py
@@ -4,16 +4,141 @@ Plot Support
 
 .. autosummary::
 
+   ~plotxy
    ~select_mpl_figure
    ~select_live_plot
    ~trim_plot_lines
    ~trim_plot_by_name
 """
 
+import datetime
 import logging
-
+import math
 
 logger = logging.getLogger(__name__)
+
+
+def plotxy(runs, xname, yname, append=False, cat=None, stats=True, stream="primary", title=None):
+    """
+    Plot y vs x from a bluesky run.
+
+    Note: This is not a bluesky plan.  Call it as a normal Python function.
+
+    PARAMETERS
+
+    ``runs`` : *[run]* or *run*:
+        List or runs or single ``run``.  A ``run`` is either a
+        ``bluesky.core.BlueskyRun`` object or a reference (uid, scan_id,
+        relative to most recent) to a BlueskyRun in the catalog.
+    ``xname`` : *str*:
+        Name of the signal to plot on the **x** axis.
+    ``yname`` : *str*:
+        Name of the signal to plot on the **y** axis.
+    ``append`` : *bool*:
+        (optional) If ``True``, append to existing plot window.
+        Default: ``append=False``
+    ``cat`` : *object*:
+        (optional) Catalog to be used for finding a run by reference.
+        Default: return value from ``apstools.utils.getCatalog()``
+    ``stats`` : *bool*:
+        (optional) If ``True``, compute and plot centroid and FWHM
+        (computed from sigma).
+        Default: ``stats=True``
+    ``stream`` : *str*:
+        (optional) Name of the data stream in which to find "xname"
+        and "yname".
+        Default: ``stream="primary"``
+    ``title`` : *str*:
+        (optional) Title to show on this plot.
+        Default: Metadata "title" keyword of first run (if found)
+        or scan_id and starting date/time of first run.
+
+    RETURNS
+
+    Returns a *dict* of statistics for each run indexed by ``scan_id``,
+    if ``stats=True``, else ``None``.  A computed ``fwhm`` key is added
+    to the statistics.
+
+    New in release 1.6.10.
+    """
+    from . import getDefaultCatalog
+    import matplotlib.pyplot as plt
+
+    plt.ion()
+
+    if not isinstance(runs, (list, tuple, range)):
+        runs = [runs]
+
+    fig_name = f"plotxy: {yname} v {xname}"
+    plt.figure(fig_name)
+    if not append:
+        plt.autoscale()
+        plt.cla()  # reset to autoscale both x & y
+
+    statistics = {}
+    for i, run in enumerate(runs):
+        if isinstance(run, (str, int)):
+            cat = cat or getDefaultCatalog()
+            run = cat.v2[run]
+
+        md = run.metadata
+        scan_id = md["start"]["scan_id"]
+        plan_name = md["start"].get("plan_name", "")
+
+        dataset = getattr(run, stream).read()
+        x = dataset[xname]
+        y = dataset[yname]
+        plt.plot(x.values, y.values, label=f"#{scan_id}: {plan_name}")
+
+        dt = datetime.datetime.fromtimestamp(md["start"]["time"])
+        if i == 0:
+            if title is None:
+                title = md["start"].get("title")
+            plt.title(f"#{scan_id}: {dt.isoformat(sep=' ')}")
+            if title is not None:
+                plt.suptitle(title)
+            plt.xlabel(x.name)
+            plt.ylabel(y.name)
+
+        if stats:
+            import pysumreg
+
+            # collect peak (and other) statistics
+            # fmt: off
+            sr = pysumreg.SummationRegisters()
+            for xv, yv, in zip(x.values, y.values):
+                sr.add(xv, yv)
+            # fmt: on
+
+            centroid = sr.centroid
+            sigma = sr.sigma
+            half_max = (sr.max_y if sr.mean_y > 0 else sr.min_y) / 2
+            # https://brainder.org/2011/08/20/gaussian-kernels-convert-fwhm-to-sigma/
+            fwhm = 2 * math.sqrt(2 * math.log(2)) * sigma
+            plt.plot(
+                [centroid] * 2, [sr.min_y, sr.max_y], color="gray", linestyle=":", label=f"_{scan_id}_centroid"
+            )
+            plt.plot(
+                [centroid - fwhm / 2, centroid + fwhm / 2],
+                [half_max] * 2,
+                color="gray",
+                linestyle=":",
+                label=f"_{scan_id}_sigma",
+            )
+            statistics[scan_id] = sr.to_dict()
+            statistics[scan_id]["fwhm"] = fwhm  # add to dictionary
+
+    plt.legend()
+    plt.text(
+        plt.axis()[1],
+        plt.axis()[2],  # inside lower-right corner
+        datetime.datetime.now(),
+        fontsize="xx-small",
+        color="gray",
+        horizontalalignment="right",
+    )
+    if len(statistics):
+        return statistics
 
 
 def select_mpl_figure(x, y):
@@ -193,6 +318,7 @@ def trim_plot_by_name(n=3, plots=None):
                     ax.lines[0].remove()
                 # update the plot legend
                 ax.legend()
+
 
 # -----------------------------------------------------------------------------
 # :author:    Pete R. Jemian